+++ udata/core/post/api.py
@@ -15,6 +15,8 @@ from udata.core.user.api_fields import user_ref_fields
 from .forms import PostForm
 from .models import Post
 
+DEFAULT_SORTING = "-published"
+
 ns = api.namespace("posts", "Posts related operations")
 
 post_fields = api.model(
@@ -58,15 +60,16 @@ post_page_fields = api.model("PostPage", fields.pager(post_fields))
 
 parser = api.page_parser()
 
-parser.add_argument(
-    "sort", type=str, default="-created_at", location="args", help="The sorting attribute"
-)
+parser.add_argument("sort", type=str, location="args", help="The sorting attribute")
 parser.add_argument(
     "with_drafts",
     type=bool,
     default=False,
     location="args",
     help="`True` also returns the unpublished posts (only for super-admins)",
+)
+parser.add_argument(
+    "q", type=str, location="args", help="query string to search through resources titles"
 )
 
 
@@ -84,7 +87,12 @@ class PostsAPI(API):
         if not (AdminPermission().can() and args["with_drafts"]):
             posts = posts.published()
 
-        return posts.order_by(args["sort"]).paginate(args["page"], args["page_size"])
+        if args["q"]:
+            phrase_query = " ".join([f'"{elem}"' for elem in args["q"].split(" ")])
+            posts = posts.search_text(phrase_query)
+
+        sort = args["sort"] or ("$text_score" if args["q"] else None) or DEFAULT_SORTING
+        return posts.order_by(sort).paginate(args["page"], args["page_size"])
 
     @api.doc("create_post")
     @api.secure(admin_permission)