+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Current (in progress)
 
 - Harvest attribution roles [#3149](https://github.com/opendatateam/udata/pull/3149)
+- Add search to posts [#3248](https://github.com/opendatateam/udata/pull/3248)
 - Resource catalog: fix preview_url and add extras [#3188](https://github.com/opendatateam/udata/pull/3188)
 - Add trailing slash to security routes [#3251](https://github.com/opendatateam/udata/pull/3251)
+- Upgrade packaging dependency to 24.2 [#3252](https://github.com/opendatateam/udata/pull/3252)
 
 ## 10.0.7 (2025-01-13)